requirements.txt
discord.py
discord.py[voice]
python-dotenv
PyNaCl
https://github.com/ytdl-org/youtube-dl/archive/refs/heads/master.zip
ffmpeg